docker: Provide TFTP service and access to serial tty
Set the required capabilities in order to allow the TFTP daemon running
as a normal (unprivileged) user.

Similarly, setup busybox to be able to launch its syslogd implementation
in order to offer access to the TFTP log messages.

Update docker.sh CLI to allow specifying a serial device path via the
'run' command. This also handles the startup of the above mentioned
daemons.

Signed-off-by: Cristian Ciocaltea <cristian.ciocaltea@gmail.com>